requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Pillow
python-dotenv==1.0.0
gunicorn==21.2.0
boto3==1.34.0
mysqlclient